requirements.txt
fastapi
h11
httptools
psycopg2-binary
pydantic
SQLAlchemy
starlette
typing
uvicorn
uvloop
websockets
transformers
protobuf
sentencepiece
python-dotenv
google-api-python-client
openai
youtube-transcript-api